refactor: Extract parsing-related code from Network class
- Extracted parsing-related code from the Network classes.
- Created new Parser classes to handle parsing logic.
- This change anticipates future additions for parsing live streams and other details.
